change all the routes to api routes
