+++ client/src/components/login.jsx
@@ -20,7 +20,7 @@ class Login extends Component {
   }
 
   sendLogin(e) {
-    axios.post('/login', this.state)
+    axios.post('/api/login', this.state)
       .then((data) => {
         window.location.replace(window.location.origin);
       })
