+++ client/src/index.jsx
@@ -7,6 +7,7 @@ import Login from './components/login.jsx';
 import Signup from './components/signup.jsx';
 import Landing from './components/landing.jsx'
 import Locations from './components/locations.jsx';
+import Upload from './components/upload.jsx';
 
 // App component renders components based on the URL Route using React Router
 const App = (props) => {
@@ -18,6 +19,7 @@ const App = (props) => {
         <Route path='/login' component={Login} />
         <Route path='/signup' component={Signup} />
         <Route path='/locations' component={Locations} />
+        <Route path='/upload' component = {Upload} />
       </Switch>
     </div>
   );
