+++ server/routes/routes.js
@@ -1,16 +1,16 @@
 const router = require('express').Router();
 const controller = require('../controller/controller.js');
 
-router.post('/signup', controller.post.signup);
+router.post('/api/signup', controller.post.signup);
 
-router.post('/login', controller.post.login);
+router.post('/api/login', controller.post.login);
 
-router.get('/logout', controller.get.logout);
+router.get('/api/logout', controller.get.logout);
 
-router.get('/loggedInYet', controller.get.loggedInYet);
+router.get('/api/loggedInYet', controller.get.loggedInYet);
 
-router.post('/upload', controller.post.upload);
+router.post('/api/upload', controller.post.upload);
 
-router.get('/upload', controller.get.upload);
+router.get('/api/upload', controller.get.upload);
 
 module.exports = router;
